modularize authconfig and passportConfig into authConfig.js
