+++ server/server.js
@@ -5,14 +5,8 @@ var cookieParser = require('cookie-parser');
 var morgan = require('morgan');
 var session = require('express-session');
 var passport = require('passport');
-var FacebookStrategy = require('passport-facebook').Strategy;
-var TwitterStrategy = require('passport-twitter').Strategy;
-var APIKeys = require('./config');
-var profileFields = require('./authConfig').profileFields;
-var options = require('./authConfig').options;
-//below are peer dependencies. They will allow you to make server side http requests
-var request = require('request');
-var rp = require('request-promise');
+//passport configuration
+var passportConfig = require('./authConfig').passportConfig;
 
 var app = express();
 
@@ -31,62 +25,7 @@ app.use(session({
 app.use(passport.initialize());
 app.use(passport.session());
 
-
-passport.use(new FacebookStrategy({
-  clientID: APIKeys.keys.facebook.key,
-  clientSecret: APIKeys.keys.facebook.secret,
-  callbackURL: 'http://localhost:3010/auth/facebook/callback',
-  profileFields: profileFields,
-},
-  function(accessToken, refreshToken, profile, done) {
-    var apiFields = options(profile.id, profile.displayName)
-
-    rp(apiFields)
-    .then(function (user) {
-        console.log(user, 'IS THIS THE DATA');
-        done(null, user);
-    })
-    .catch(function (err) {
-      console.log(err,'could not reach SquirrelDBService');
-    });
-  }
-));
-
-passport.use(new TwitterStrategy({
-  consumerKey: APIKeys.keys.twitter.key,
-  consumerSecret: APIKeys.keys.twitter.secret,
-  callbackURL: 'http://localhost:3010/auth/twitter/callback'
-},
-function(token, tokenSecret, profile, done) {
-  console.log(profile, 'twitter profile?');
-  User.findOrCreate({}, function(err, user) {
-    if (err) {
-      return done(err);
-    }
-    done(null, user);
-  });
-}));
-
-//user ID is serialized to the session, when a request of the same ID is received it will restore the session
-passport.serializeUser(function(user, done) {
-  done(null, user.fbid);
-});
-//used to check if user session is actuallly a verified user in database! 
-passport.deserializeUser(function(id, done) {
-  var apiFields = options(id);
-
-  rp(apiFields)
-    .then(function(user){
-      console.log('passport.deserilizeUser', user);
-      done(err, user);
-    })
-    .catch(function(err){
-      console.log('passport.deserilizeUser 2', err);
-    })
-
-});
-
-//app.get('/');
+passportConfig(passport);
 
 app.post('/login', passport.authenticate('local'), function(req, res) {
   //if this function gets invoked, authentucation was successful
